feat: refactoring and closing 11 homework

+++ src/features/comments/service/commentsService.ts
@@ -1,22 +1,57 @@
 import { CommentsRepositoryImpl } from '../database/commentsRepoImpl';
-import { UserViewModel } from '../../users/models/User';
 import { createResult, Result } from '../../../shared/utils/result-object';
 import { inject, injectable } from 'inversify';
+import { AuthDocument } from '../../auth/database/authEntity';
+import { LikeService } from '../../like/likeService';
+import { CommentModelEntity } from '../database/commentsEntity';
 
 @injectable()
 export class CommentsService {
   constructor(
-    @inject(CommentsRepositoryImpl) readonly commentsRepository: CommentsRepositoryImpl
+    @inject(CommentsRepositoryImpl) readonly commentsRepository: CommentsRepositoryImpl,
+    @inject(LikeService) readonly likesService: LikeService
   ) {}
+  async updateComment(
+    id: string,
+    content: string,
+    userId: string
+  ): Promise<Result<boolean | null>> {
+    const comment = await this.commentsRepository.getCommentById(id);
+    if (!comment) return createResult('NOT_FOUND', null);
 
-  async updateComment(id: string, content: string): Promise<Result<boolean>> {
-    const resultUpdate = await this.commentsRepository.updateComment(id, content);
-    if (!resultUpdate) return createResult('NOT_FOUND', resultUpdate);
+    if (comment.commentatorInfo.userId !== userId) return createResult('FORBIDDEN', null);
 
-    return createResult('NO_CONTENT', resultUpdate);
+    // const isUserComment = await this.commentsRepository.getCommentByUserIdAndCommentId(id, userId);
+    // if (!isUserComment) return createResult('FORBIDDEN', null);
+
+    comment.content = content;
+    await this.commentsRepository.save(comment);
+
+    return createResult('NO_CONTENT', true);
   }
 
-  async deleteComment(id: string): Promise<Result<boolean>> {
+  async updateCommentLikesInfo(
+    id: string,
+    likesInfo: { likesCount: number; dislikesCount: number }
+  ): Promise<Result<boolean>> {
+    const comment = await this.commentsRepository.getCommentById(id);
+    if (!comment) return createResult('NOT_FOUND', false);
+
+    // const resultUpdate = await this.commentsRepository.updateCommentLikesInfo(id, likesInfo);
+    // if (!resultUpdate) return createResult('NOT_FOUND', resultUpdate);
+    comment.likesInfo.likesCount = likesInfo.likesCount;
+    comment.likesInfo.dislikesCount = likesInfo.dislikesCount;
+    await this.commentsRepository.save(comment);
+
+    return createResult('NO_CONTENT', true);
+  }
+
+  async deleteComment(id: string, userId: string): Promise<Result<boolean | null>> {
+    const comment = await this.commentsRepository.getCommentById(id);
+    if (!comment) return createResult('NOT_FOUND', null);
+
+    if (comment.commentatorInfo.userId !== userId) return createResult('FORBIDDEN', null);
+
     const resultDelete = await this.commentsRepository.deleteComment(id);
     if (!resultDelete) return createResult('NOT_FOUND', resultDelete);
 
@@ -26,20 +61,40 @@ export class CommentsService {
   async createCommentByPostId(
     postId: string,
     content: string,
-    user: UserViewModel
+    user: AuthDocument
   ): Promise<Result<string | null>> {
-    const comment = {
+    const comment = new CommentModelEntity({
       postId,
       content,
       commentatorInfo: {
         userId: user.id,
         userLogin: user.login,
       },
-      createdAt: new Date(),
-    };
-    const result = await this.commentsRepository.createCommentByPostId(comment);
-    if (!result) return createResult('BAD_REQUEST', result);
+    });
 
-    return createResult('CREATED', result);
+    await this.commentsRepository.save(comment);
+    return createResult('CREATED', comment.id);
+  }
+
+  async changeLikeStatus(
+    commentId: string,
+    userId: string,
+    likeStatus: string
+  ): Promise<Result<{ likesCount: number; dislikesCount: number } | null>> {
+    const comment = await this.commentsRepository.getCommentById(commentId);
+    if (!comment) return createResult('NOT_FOUND', null);
+
+    const resultManageLike = await this.likesService.manageLike(
+      userId,
+      commentId,
+      likeStatus,
+      comment.postId
+    );
+
+    if (resultManageLike.data !== null) {
+      await this.updateCommentLikesInfo(commentId, resultManageLike.data);
+    }
+
+    return resultManageLike;
   }
 }
